requirements.txt
marker-pdf
requests
